// src/apiClient/models/updateWordEntityRequest.ts
// tslint:disable
// eslint-disable
// Generated by Microsoft Kiota
import { type Parsable, type ParseNode, type SerializationWriter } from '@microsoft/kiota-abstractions';
import { type Guid } from 'guid-typescript';

export function createUpdateWordEntityRequestFromDiscriminatorValue(parseNode: ParseNode | undefined) {
    if(!parseNode) throw new Error("parseNode cannot be undefined");
    return deserializeIntoUpdateWordEntityRequest;
}
export function deserializeIntoUpdateWordEntityRequest(updateWordEntityRequest: UpdateWordEntityRequest | undefined = {} as UpdateWordEntityRequest) : Record<string, (node: ParseNode) => void> {
    return {
        "id": n => { updateWordEntityRequest.id = n.getGuidValue(); },
        "language": n => { updateWordEntityRequest.language = n.getNumberValue(); },
        "status": n => { updateWordEntityRequest.status = n.getNumberValue(); },
        "word": n => { updateWordEntityRequest.word = n.getStringValue(); },
    }
}
export function serializeUpdateWordEntityRequest(writer: SerializationWriter, updateWordEntityRequest: UpdateWordEntityRequest | undefined = {} as UpdateWordEntityRequest) : void {
        writer.writeGuidValue("id", updateWordEntityRequest.id);
        writer.writeNumberValue("language", updateWordEntityRequest.language);
        writer.writeNumberValue("status", updateWordEntityRequest.status);
        writer.writeStringValue("word", updateWordEntityRequest.word);
}
export interface UpdateWordEntityRequest extends Parsable {
    /**
     * The id property
     */
    id?: Guid;
    /**
     * The language property
     */
    language?: number;
    /**
     * The status property
     */
    status?: number;
    /**
     * The word property
     */
    word?: string;
}
// tslint:enable
// eslint-enable